% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_list.R
\name{addNewLib}
\alias{addNewLib}
\title{Add a new library}
\usage{
addNewLib(newlib)
}
\arguments{
\item{newlib}{A path to a valid directory where R has installed packages}
}
\description{
Add the new library to the R .libPaths() list.
}

