% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_list.R
\name{install_pkgs}
\alias{install_pkgs}
\title{Restore Packages from a Backup}
\usage{
install_pkgs(pkgList, libpath = NULL, update = FALSE, keep_all = TRUE)
}
\arguments{
\item{pkgList}{An R vector of package names, ideally read in from \code{\link{read_pkgs}}}

\item{libpath}{A path to a valid directory where R can install the packages}

\item{update}{A logical, should packages be updated first? Default is false}

\item{keep_all}{A logical, should all packages in pkgList not local be installed?}

\item{...}{additional arguments to pass to \code{\link{sync_pkgs}}}
}
\description{
Read package names from a CSV file to an R object.
}

