% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintainr-package.r
\docType{package}
\name{maintainr}
\alias{maintainr}
\alias{maintainr-package}
\title{maintainr: Provides a simple method for saving and restoring a list of R
packages installed on a machine.}
\description{
maintainr is not intended to be a replacement for more comprehensive archiving
packages such as packrat or MRAN, but instead is just a simple way to save
a list of installed packages to a central location and read that list into
a new R install and optionally install the packages.
}

