% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudUploaders.R
\name{pkg_restore}
\alias{pkg_restore}
\title{Restore R packages from a package list stored somewhere else}
\usage{
pkg_restore(install = TRUE, backupPath = NULL, filename = NULL,
  cloudProvider = NULL, libpath = NULL, update = FALSE,
  keep_all = FALSE, ...)
}
\arguments{
\item{install}{logical, should packages in the package list be installed?}

\item{backupPath}{The path to where \code{\link{savePkgs}} should write the 
list of packages, optional}

\item{filename}{The name of the csv file to save the package names, optional}

\item{cloudProvider}{Optional, if specified the package will attempt to write 
the package list to a tempfile and then upload that tempfile to a cloud service}

\item{libpath}{Path to the library}

\item{update}{A logical, should packages be updated first? Default is false}

\item{keep_all}{A logical, should all packages in pkgList not local be installed?}

\item{...}{Additional arguments to pass to \code{\link{readCloud}}}
}
\value{
Nothing.
}
\description{
Restore R packages from a package list stored somewhere else
}

