% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudUploaders.R
\name{read_cloud}
\alias{read_cloud}
\title{Cloud fetch handler}
\usage{
read_cloud(filename, provider = c("dropbox"), ...)
}
\arguments{
\item{filename}{path to file to be uploaded to cloud storage}

\item{provider}{a character for the provider to use, currently only "dropbox"
is supported}

\item{...}{additional arguments to pass to the function that uploads the file}
}
\value{
Nothing.
}
\description{
Cloud fetch handler
}
\seealso{
\code{\link[rdrop2:drop_read_csv]{drop_upload}}
}

