\name{readPkgs}
\alias{readPkgs}
\title{Read Packages}
\usage{
  readPkgs(path, filename)
}
\arguments{
  \item{path}{A string that lists a valid path for R to the
  place to import the list of packages. Ideally to a remote
  backup solution.}
}
\description{
  Read package names from a CSV file to an R object.
}

