\name{save_pkgs}
\alias{save_pkgs}
\title{Save Packages}
\usage{
  save_pkgs(path)
}
\arguments{
  \item{path}{A string that lists a valid path for R to the
  place to export the list of packages. Ideally to a remote
  backup solution.}
}
\description{
  Export package names to a CSV file to be restored later.
}

