% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_list.R
\name{setLibrary}
\alias{setLibrary}
\title{Set the library}
\usage{
setLibrary()
}
\description{
Tell R where to find new packages in a new library in the current session.
}
\author{
Tal Galili
}
\seealso{
\url{http://www.r-statistics.com/2010/04/changing-your-r-upgrading-strategy-and-the-r-code-to-do-it-on-windows/}
}

