% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_list.R
\name{sync_pkgs}
\alias{sync_pkgs}
\title{Check versions to install only newer packages}
\usage{
sync_pkgs(pkgList, keep_all = TRUE)
}
\arguments{
\item{pkgList}{A new list of packages}

\item{keep_all}{logical, should all packages in the new list that are not 
installed on the system get synced up}
}
\value{
A list of packages to be installed that are not installed or out of date
}
\description{
Check versions to install only newer packages
}

