% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudUploaders.R
\name{writeCloud}
\alias{writeCloud}
\title{Cloud upload handler}
\usage{
writeCloud(filename, provider = c("dropbox"), ...)
}
\arguments{
\item{filename}{path and filename of file to be uploaded to cloud storage}

\item{provider}{a character for the provider to use, currently only "dropbox"
is supported}

\item{...}{additional arguments to pass to the function that uploads the file}
}
\value{
Nothing.
}
\description{
Cloud upload handler
}
\seealso{
\code{\link[rdrop2:drop_read_csv]{drop_upload}}
}

