% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_list.R
\name{readPkgs}
\alias{readPkgs}
\title{Read Packages}
\usage{
readPkgs(path, filename)
}
\arguments{
\item{path}{A string that lists a valid path for R to the place to import the list}

\item{filename}{A string that denotes the filename where the packages are stored
of packages. Ideally to a remote backup solution.}
}
\description{
Read package names from a CSV file to an R object.
}

