% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_profile.R
\name{writeProfile}
\alias{writeProfile}
\title{Write a new .Rprofile}
\usage{
writeProfile(pkglist)
}
\arguments{
\item{pkglist}{A pkglist produced by \code{\link{savePkgs}}}
}
\description{
Export a custom .Rprofile for syncing between installs.
}
\seealso{
\url{http://stackoverflow.com/questions/1401904/painless-way-to-install-a-new-version-of-r-on-windows}
}

